requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
cloudinary==1.42.1
dnspython==1.16.0
exceptiongroup==1.2.2
fastapi==0.115.7
h11==0.14.0
httptools==0.6.4
idna==3.10
motor==3.6.1
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.9.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
